+++ src/main/java/com/bvk/partner/tokopedia/seller/object/RequestPickup.java
@@ -2,8 +2,8 @@ package com.bvk.partner.tokopedia.seller.object;
 
 public class RequestPickup {
 	
-	public Integer order_id;
-    public Integer shop_id;
+	public Long order_id;
+    public Long shop_id;
     public String request_time;
     public String result;
