+++ src/main/java/com/bvk/partner/tokopedia/seller/object/UploadStatus.java
@@ -7,7 +7,7 @@ public class UploadStatus {
 	public ArrayList<UploadDatum> upload_data;
 	
 	public class UploadDatum {
-	    public Integer upload_id;
+	    public Long upload_id;
 	    public String status;
 	    public Integer total_data;
 	    public Integer unprocessed_rows;
